Cleared up SqlServerUserRepository

+++ back-end/DataMigrationApi.Services/SqlServerUserService.cs
@@ -1,6 +1,7 @@
 ﻿using DataMigrationApi.Core.Abstractions;
 using DataMigrationApi.Core.Abstractions.Services;
 using DataMigrationApi.Core.Entities;
+using System;
 using System.Collections.Generic;
 
 namespace DataMigrationApi.Services
@@ -26,22 +27,40 @@ namespace DataMigrationApi.Services
             return user;
         }
 
-        public User Insert(User user)
+        public User Insert(User entity)
         {
-            _unitOfWork.SqlServerUserRepository.Insert(user);
+            if (!Guid.TryParse(entity.ID, out _) || Get(entity.ID) != null)
+            {
+                entity.ID = Guid.NewGuid().ToString();
+            }
+
+            _unitOfWork.SqlServerUserRepository.Insert(entity);
             _unitOfWork.Save();
-            return user;
+            return entity;
         }
 
-        public User Update(User user)
+        public User Update(User entity)
         {
-            var userToUpdate = _unitOfWork.SqlServerUserRepository.Update(user);
+            var user = Get(entity.ID);
+            if (user == null)
+            {
+                return null;
+            }
+
+            entity.Identity = user.Identity;
+            var userToUpdate = _unitOfWork.SqlServerUserRepository.Update(entity);
+
             _unitOfWork.Save();
             return userToUpdate;
         }
 
         public void Delete(string id)
         {
+            var user = Get(id);
+            if (user == null)
+            {
+                return;
+            }
             _unitOfWork.SqlServerUserRepository.Delete(id);
             _unitOfWork.Save();
         }
